Move all of the infer code to their own files
