+++ test/basic.js
@@ -3,14 +3,12 @@
 const common = require('../common')
 const config = require('./config.json')
 const fs = require('fs-extra')
-const os = require('os')
 const packager = require('..')
 const path = require('path')
 const targets = require('../targets')
 const test = require('tape')
 const util = require('./util')
 const waterfall = require('run-waterfall')
-const pkgUp = require('pkg-up')
 
 // Generates a path to the generated app that reflects the name given in the options.
 // Returns the Helper.app location on darwin since the top-level .app is already tested for the
@@ -205,101 +203,6 @@ function createOverwriteTest (opts) {
   }
 }
 
-function createInferElectronVersionTest (fixture, packageName) {
-  return (opts) => {
-    return (t) => {
-      t.timeoutAfter(config.timeout)
-
-      // Don't specify name or version
-      delete opts.electronVersion
-      opts.dir = path.join(__dirname, 'fixtures', fixture)
-
-      waterfall([
-        (cb) => {
-          common.getMetadataFromPackageJSON(opts, opts.dir, cb)
-        }, (cb) => {
-          fs.readFile(path.join(opts.dir, 'package.json'), cb)
-        }, (pkg, cb) => {
-          const packageJSON = JSON.parse(pkg)
-          t.equal(opts.electronVersion, packageJSON.devDependencies[packageName], `The version should be inferred from installed ${packageName} version`)
-          cb()
-        }
-      ], (err) => {
-        t.end(err)
-      })
-    }
-  }
-}
-
-function copyFixtureToTempDir (fixtureSubdir, cb) {
-  let tmpdir = path.join(os.tmpdir(), fixtureSubdir)
-  let fixtureDir = path.join(__dirname, 'fixtures', fixtureSubdir)
-  waterfall([
-    cb => {
-      let tmpdirPkg = pkgUp.sync(path.join(tmpdir, '..'))
-      if (tmpdirPkg) return cb(new Error(`Found package.json in parent of temp directory, which will interfere with test results. Please remove package.json at ${tmpdirPkg}`))
-      cb()
-    },
-    cb => fs.emptyDir(tmpdir, cb),
-    (cb1, cb2) => fs.copy(fixtureDir, tmpdir, cb2 || cb1), // inconsistent cb arguments from fs.emptyDir
-    cb => cb(null, tmpdir)
-  ], cb)
-}
-
-function createInferFailureTest (opts, fixtureSubdir) {
-  return function (t) {
-    t.timeoutAfter(config.timeout)
-
-    copyFixtureToTempDir(fixtureSubdir, (err, dir) => {
-      if (err) return t.end(err)
-
-      delete opts.electronVersion
-      opts.dir = dir
-
-      packager(opts, function (err, paths) {
-        t.ok(err, 'error thrown')
-        t.end()
-      })
-    })
-  }
-}
-
-function createInferMissingVersionTest (opts) {
-  return (t) => {
-    t.timeoutAfter(config.timeout)
-    waterfall([
-      (cb) => {
-        copyFixtureToTempDir('infer-missing-version-only', cb)
-      }, (dir, cb) => {
-        delete opts.electronVersion
-        opts.dir = dir
-
-        common.getMetadataFromPackageJSON(opts, dir, cb)
-      }, (cb) => {
-        fs.readFile(path.join(opts.dir, 'package.json'), cb)
-      }, (pkg, cb) => {
-        const packageJSON = JSON.parse(pkg)
-        t.equal(opts.electronVersion, packageJSON.devDependencies['electron'], 'The version should be inferred from installed electron module version')
-        cb()
-      }
-    ], (err) => {
-      t.end(err)
-    })
-  }
-}
-
-function createInferMissingFieldsTest (opts) {
-  return createInferFailureTest(opts, 'infer-missing-fields')
-}
-
-function createInferWithBadFieldsTest (opts) {
-  return createInferFailureTest(opts, 'infer-bad-fields')
-}
-
-function createInferWithMalformedJSONTest (opts) {
-  return createInferFailureTest(opts, 'infer-malformed-json')
-}
-
 function createTmpdirTest (opts) {
   return function (t) {
     t.timeoutAfter(config.timeout)
@@ -465,12 +368,6 @@ test('cannot build apps where the name ends in " Helper"', (t) => {
   })
 })
 
-util.testSinglePlatform('infer test using `electron-prebuilt` package', createInferElectronVersionTest('basic', 'electron-prebuilt'))
-util.testSinglePlatform('infer test using `electron` package only', createInferMissingVersionTest)
-util.testSinglePlatform('infer test where `electron` version is preferred over `electron-prebuilt`', createInferElectronVersionTest('basic-renamed-to-electron', 'electron'))
-util.testSinglePlatform('infer missing fields test', createInferMissingFieldsTest)
-util.testSinglePlatform('infer with bad fields test', createInferWithBadFieldsTest)
-util.testSinglePlatform('infer with malformed JSON test', createInferWithMalformedJSONTest)
 util.testSinglePlatform('defaults test', createDefaultsTest)
 util.testSinglePlatform('out test', createOutTest)
 util.testSinglePlatform('prune test', (baseOpts) => {
